Add write mesh abstraction. TODO: Test for h5py
