+++ src/adios4dolfinx/backends/adios2/backend.py
@@ -1,3 +1,4 @@
+import warnings
 from pathlib import Path
 from typing import Any, Union
 
@@ -7,7 +8,9 @@ import adios2
 import numpy as np
 import numpy.typing as npt
 
-from adios4dolfinx.utils import FileMode, check_file_exists
+from adios4dolfinx.interface import FileMode
+from adios4dolfinx.structures import MeshData
+from adios4dolfinx.utils import check_file_exists
 
 from .helpers import ADIOSFile, adios_to_numpy_dtype, resolve_adios_scope
 
@@ -151,3 +154,108 @@ class ADIOS2Interface:
                 adios_file.file.EndStep()
 
         return np.array(time_stamps)
+
+    @staticmethod
+    def write_mesh(
+        filename: Path,
+        comm: MPI.Intracomm,
+        mesh: MeshData,
+        backend_args: dict[str, Any] | None = None,
+        mode: FileMode = FileMode.write,
+        time: float = 0.0,
+    ):
+        """
+        Write a mesh to file using ADIOS2
+
+        Parameters:
+            comm: MPI communicator used in storage
+            mesh: Internal data structure for the mesh data to save to file
+            filename: Path to file to write to
+            engine: ADIOS2 engine to use
+            mode: ADIOS2 mode to use (write or append)
+            io_name: Internal name used for the ADIOS IO object
+        """
+        if "io_name" not in backend_args.keys():
+            io_name = "MeshWriter"
+        else:
+            io_name = backend_args["io_name"]
+        engine = backend_args["engine"]
+
+        gdim = mesh.local_geometry.shape[1]
+        adios = adios2.ADIOS(comm)
+        with ADIOSFile(
+            adios=adios, filename=filename, mode=mode, engine=engine, io_name=io_name, comm=comm
+        ) as adios_file:
+            adios_file.file.BeginStep()
+            # Write geometry
+            pointvar = adios_file.io.DefineVariable(
+                "Points",
+                mesh.local_geometry,
+                shape=[mesh.num_nodes_global, gdim],
+                start=[mesh.local_geometry_pos[0], 0],
+                count=[mesh.local_geometry_pos[1] - mesh.local_geometry_pos[0], gdim],
+            )
+            adios_file.file.Put(pointvar, mesh.local_geometry, adios2.Mode.Sync)
+
+            if mode == adios2.Mode.Write:
+                adios_file.io.DefineAttribute("CellType", mesh.cell_type)
+                adios_file.io.DefineAttribute("Degree", np.array([mesh.degree], dtype=np.int32))
+                adios_file.io.DefineAttribute(
+                    "LagrangeVariant", np.array([mesh.lagrange_variant], dtype=np.int32)
+                )
+                # Write topology (on;y on first write as topology is constant)
+                num_dofs_per_cell = mesh.local_topology.shape[1]
+                dvar = adios_file.io.DefineVariable(
+                    "Topology",
+                    mesh.local_topology,
+                    shape=[mesh.num_cells_global, num_dofs_per_cell],
+                    start=[mesh.local_topology_pos[0], 0],
+                    count=[
+                        mesh.local_topology_pos[1] - mesh.local_topology_pos[0],
+                        num_dofs_per_cell,
+                    ],
+                )
+                adios_file.file.Put(dvar, mesh.local_topology)
+
+                # Add partitioning data
+                if mesh.store_partition:
+                    assert mesh.partition_range is not None
+                    par_data = adios_file.io.DefineVariable(
+                        "PartitioningData",
+                        mesh.ownership_array,
+                        shape=[mesh.partition_global],
+                        start=[mesh.partition_range[0]],
+                        count=[
+                            mesh.partition_range[1] - mesh.partition_range[0],
+                        ],
+                    )
+                    adios_file.file.Put(par_data, mesh.ownership_array)
+                    assert mesh.ownership_offset is not None
+                    par_offset = adios_file.io.DefineVariable(
+                        "PartitioningOffset",
+                        mesh.ownership_offset,
+                        shape=[mesh.num_cells_global + 1],
+                        start=[mesh.local_topology_pos[0]],
+                        count=[mesh.local_topology_pos[1] - mesh.local_topology_pos[0] + 1],
+                    )
+                    adios_file.file.Put(par_offset, mesh.ownership_offset)
+                    assert mesh.partition_processes is not None
+                    adios_file.io.DefineAttribute(
+                        "PartitionProcesses", np.array([mesh.partition_processes], dtype=np.int32)
+                    )
+            if mode == adios2.Mode.Append and mesh.store_partition:
+                warnings.warn("Partitioning data is not written in append mode")
+
+            # Add time step to file
+            t_arr = np.array([time], dtype=np.float64)
+            time_var = adios_file.io.DefineVariable(
+                "MeshTime",
+                t_arr,
+                shape=[1],
+                start=[0],
+                count=[1 if comm.rank == 0 else 0],
+            )
+            adios_file.file.Put(time_var, t_arr)
+
+            adios_file.file.PerformPuts()
+            adios_file.file.EndStep()
