+++ src/adios4dolfinx/writers.py
@@ -1,33 +1,123 @@
-# Copyright (C) 2024 Jørgen Schartum Dokken
+# Copyright (C) 2024-2026 Jørgen Schartum Dokken
 #
 # This file is part of adios4dolfinx
 #
 # SPDX-License-Identifier:    MIT
 
-
-import warnings
 from pathlib import Path
+from typing import Any, Literal
 
 from mpi4py import MPI
 
 import adios2
+import dolfinx
 import numpy as np
+from packaging.version import Version
+
+from adios4dolfinx.interface import FileMode, get_backend
 
 from .backends.adios2.helpers import ADIOSFile, check_variable_exists, resolve_adios_scope
 from .structures import FunctionData, MeshData
-from .utils import FileMode
 
 adios2 = resolve_adios_scope(adios2)
+
+
+def prepare_meshdata_for_storage(mesh: dolfinx.mesh.Mesh, store_partition_info: bool) -> MeshData:
+    """
+    Helper function for extracting the required data from a distributed
+    {py:class}`dolfinx.mesh.Mesh`.
+
+    Args:
+        mesh: The mesh
+        store_partition_info: If one should store the partitioning info
+    Returns:
+        Data-container with the info that should be stored.
+    """
+
+    num_xdofs_local = mesh.geometry.index_map().size_local
+    num_xdofs_global = mesh.geometry.index_map().size_global
+    geometry_range = mesh.geometry.index_map().local_range
+    gdim = mesh.geometry.dim
+
+    # Convert local connectivity to globa l connectivity
+    g_imap = mesh.geometry.index_map()
+    g_dmap = mesh.geometry.dofmap
+    num_cells_local = mesh.topology.index_map(mesh.topology.dim).size_local
+    num_cells_global = mesh.topology.index_map(mesh.topology.dim).size_global
+    cell_range = mesh.topology.index_map(mesh.topology.dim).local_range
+    cmap = mesh.geometry.cmap
+    geom_layout = cmap.create_dof_layout()
+    if hasattr(geom_layout, "num_entity_closure_dofs"):
+        num_dofs_per_cell = geom_layout.num_entity_closure_dofs(mesh.topology.dim)
+    else:
+        num_dofs_per_cell = len(geom_layout.entity_closure_dofs(mesh.topology.dim, 0))
+    dofs_out = np.zeros((num_cells_local, num_dofs_per_cell), dtype=np.int64)
+    assert g_dmap.shape[1] == num_dofs_per_cell
+    dofs_out[:, :] = np.asarray(
+        g_imap.local_to_global(g_dmap[:num_cells_local, :].reshape(-1))
+    ).reshape(dofs_out.shape)
+
+    if store_partition_info:
+        partition_processes = mesh.comm.size
+
+        # Get partitioning
+        if Version(dolfinx.__version__) > Version("0.9.0"):
+            consensus_tag = 1202
+            cell_map = mesh.topology.index_map(mesh.topology.dim).index_to_dest_ranks(consensus_tag)
+        else:
+            cell_map = mesh.topology.index_map(mesh.topology.dim).index_to_dest_ranks()
+        num_cells_local = mesh.topology.index_map(mesh.topology.dim).size_local
+        cell_offsets = cell_map.offsets[: num_cells_local + 1]
+        if cell_offsets[-1] == 0:
+            cell_array = np.empty(0, dtype=np.int32)
+        else:
+            cell_array = cell_map.array[: cell_offsets[-1]]
+
+        # Compute adjacency with current process as first entry
+        ownership_array = np.full(num_cells_local + cell_offsets[-1], -1, dtype=np.int32)
+        ownership_offset = cell_offsets + np.arange(len(cell_offsets), dtype=np.int32)
+        ownership_array[ownership_offset[:-1]] = mesh.comm.rank
+        insert_position = np.flatnonzero(ownership_array == -1)
+        ownership_array[insert_position] = cell_array
+
+        partition_map = dolfinx.common.IndexMap(mesh.comm, ownership_array.size)
+        ownership_offset += partition_map.local_range[0]
+        partition_range = partition_map.local_range
+        partition_global = partition_map.size_global
+    else:
+        partition_processes = None
+        ownership_array = None
+        ownership_offset = None
+        partition_range = None
+        partition_global = None
+
+    return MeshData(
+        local_geometry=mesh.geometry.x[:num_xdofs_local, :gdim].copy(),
+        local_geometry_pos=geometry_range,
+        num_nodes_global=num_xdofs_global,
+        local_topology=dofs_out,
+        local_topology_pos=cell_range,
+        num_cells_global=num_cells_global,
+        cell_type=mesh.topology.cell_name(),
+        degree=mesh.geometry.cmap.degree,
+        lagrange_variant=mesh.geometry.cmap.variant,
+        store_partition=store_partition_info,
+        partition_processes=partition_processes,
+        ownership_array=ownership_array,
+        ownership_offset=ownership_offset,
+        partition_range=partition_range,
+        partition_global=partition_global,
+    )
 
 
 def write_mesh(
     filename: Path,
     comm: MPI.Intracomm,
-    mesh: MeshData,
-    engine: str = "BP4",
-    mode: adios2.Mode = adios2.Mode.Write,
+    mesh_data: MeshData,
+    mode: FileMode = FileMode.write,
     time: float = 0.0,
-    io_name: str = "MeshWriter",
+    backend_args: dict[str, Any] | None = None,
+    backend: Literal["adios2", "h5py"] = "adios2",
 ):
     """
     Write a mesh to file using ADIOS2
@@ -40,85 +130,11 @@ def write_mesh(
         mode: ADIOS2 mode to use (write or append)
         io_name: Internal name used for the ADIOS IO object
     """
-
-    gdim = mesh.local_geometry.shape[1]
-    adios = adios2.ADIOS(comm)
-    with ADIOSFile(
-        adios=adios, filename=filename, mode=mode, engine=engine, io_name=io_name, comm=comm
-    ) as adios_file:
-        adios_file.file.BeginStep()
-        # Write geometry
-        pointvar = adios_file.io.DefineVariable(
-            "Points",
-            mesh.local_geometry,
-            shape=[mesh.num_nodes_global, gdim],
-            start=[mesh.local_geometry_pos[0], 0],
-            count=[mesh.local_geometry_pos[1] - mesh.local_geometry_pos[0], gdim],
-        )
-        adios_file.file.Put(pointvar, mesh.local_geometry, adios2.Mode.Sync)
-
-        if mode == adios2.Mode.Write:
-            adios_file.io.DefineAttribute("CellType", mesh.cell_type)
-            adios_file.io.DefineAttribute("Degree", np.array([mesh.degree], dtype=np.int32))
-            adios_file.io.DefineAttribute(
-                "LagrangeVariant", np.array([mesh.lagrange_variant], dtype=np.int32)
-            )
-            # Write topology (on;y on first write as topology is constant)
-            num_dofs_per_cell = mesh.local_topology.shape[1]
-            dvar = adios_file.io.DefineVariable(
-                "Topology",
-                mesh.local_topology,
-                shape=[mesh.num_cells_global, num_dofs_per_cell],
-                start=[mesh.local_topology_pos[0], 0],
-                count=[
-                    mesh.local_topology_pos[1] - mesh.local_topology_pos[0],
-                    num_dofs_per_cell,
-                ],
-            )
-            adios_file.file.Put(dvar, mesh.local_topology)
-
-            # Add partitioning data
-            if mesh.store_partition:
-                assert mesh.partition_range is not None
-                par_data = adios_file.io.DefineVariable(
-                    "PartitioningData",
-                    mesh.ownership_array,
-                    shape=[mesh.partition_global],
-                    start=[mesh.partition_range[0]],
-                    count=[
-                        mesh.partition_range[1] - mesh.partition_range[0],
-                    ],
-                )
-                adios_file.file.Put(par_data, mesh.ownership_array)
-                assert mesh.ownership_offset is not None
-                par_offset = adios_file.io.DefineVariable(
-                    "PartitioningOffset",
-                    mesh.ownership_offset,
-                    shape=[mesh.num_cells_global + 1],
-                    start=[mesh.local_topology_pos[0]],
-                    count=[mesh.local_topology_pos[1] - mesh.local_topology_pos[0] + 1],
-                )
-                adios_file.file.Put(par_offset, mesh.ownership_offset)
-                assert mesh.partition_processes is not None
-                adios_file.io.DefineAttribute(
-                    "PartitionProcesses", np.array([mesh.partition_processes], dtype=np.int32)
-                )
-        if mode == adios2.Mode.Append and mesh.store_partition:
-            warnings.warn("Partitioning data is not written in append mode")
-
-        # Add time step to file
-        t_arr = np.array([time], dtype=np.float64)
-        time_var = adios_file.io.DefineVariable(
-            "MeshTime",
-            t_arr,
-            shape=[1],
-            start=[0],
-            count=[1 if comm.rank == 0 else 0],
-        )
-        adios_file.file.Put(time_var, t_arr)
-
-        adios_file.file.PerformPuts()
-        adios_file.file.EndStep()
+    backend = get_backend(backend)
+    backend_args = backend.get_default_backend_args(backend_args)
+    backend.write_mesh(
+        filename, comm, mesh_data, backend_args, backend.convert_file_mode(mode), time
+    )
 
 
 def write_function(
